OF-3059: Remove AnonymousUserCache from RoutingTable
To maintain distinction between regular and anonymous ‘routes’, the RoutingTableImpl class used different caches for each. For every use case (other than determining if a ‘route is anonymous’),  both caches are operated on. This adds a lot of code complexity, and, as caches are operated on under cluster-wide lock, resource usage.

As the 'route is anonymous' concept has been replaced in the previous commit, the only reason to maintain two caches has been removed.

This commit removes the AnonymousUserCache completely. All data that was previously stored in either UserCache or AnonymousUserCache is now stored in UserCache.