CI: Allow Openfire to log at different levels
This modifies the GitHub action that's used to start Openfire to take an optional log level. When set, the log4j configuration used by Openfire is modified accordingly.

The continuous integration jobs that use this action are now logging on level 'debug'. That should provide more analytics, in case a step fails.