DeliverRawTestTask to obtain sessions from SessionManager rather than RoutingTable
Sessions obtained from a routing table are supposed to be _route-able_. Although many/most use cases probably intend to use such a session, sessions obtained from the SessionManager can include more sessions (such as pre-authenticated sessions).

When a session is desired, that aught to be retreived from the SessionManager, not a RoutingTable.