OF-3061: Improve DB query to get last pubsub items
The old codebase defined two queries to get the last item for a pubsub node, and to get all items for a pubsub node. The actual SQL of both definitions is the same. It does not contain any ‘limit’ or ‘top’ clause (although database driver hints are given).

This is likely the result of Openfire supporting multiple database systems, that each have their own clauses to limit a resultset. SQLServer uses SELECT TOP(?) ..., Oracle has ... FETCH FIRST ? ROWS ONLY while other databases generally use ... LIMIT ?. It is inconvenient to fold all these alternatives in one implementation.

In this commit, the code is modified to:

- limit the size of the requested resultset from the database appropriately.
- remove SQL duplication (getting the latest item can use the same query as getting the latest X items, using an argument value of 1).

This intends to make the database interaction more efficient (by reducing the likelihood that the database offers up all items for a node, when only a few (or one!) is needed.