OF-3062: Improve Roster Set processing (ignore subscription value)
RFC 6121 section 2.1.5 defines:

> The following rules apply to roster sets: [...] The server MUST ignore any value of the 'subscription' attribute other than "remove"

Openfire currently ignores subscription values that are otherwise valid subscription item values (such as none and both) but it throws an error when a different value (e.g.foobar) is used.

The specification clearly states that any value MUST be ignored, so returning an error is not quite right. Instead, that value should be ignored and the Roster Set should be processed as if it wasn’t there.