OF-3065: Improve Admin Console config for Connection Managers
This commit brings the admin console pages on which Connection Manager (Multiplexer) connectivity is configured more in line with the configuration for other types of connectivity, like Clients, Servers and Components.

The biggest functional improvement that this brings is that the DirectTLS port can now be configured, and enabled/disabled without requiring a server restart.