OF-2280: Prevent closure of Server sessions of other cluster nodes
When removing an outgoing server to server session, this must be invoked from the cluster node that holds the physical connection (this is a pre-existing contract).

With this change, Openfire explicitly checks if the closing server indeed is the owner of the physical server (disregarding the request otherwise).

This intends to reduce the amount of cluster cache inconsistencies that relate to server-to-server sessions (such as: "Not all elements in of LocalServerRoutingTable's response exist in Routing Servers Cache.").