OF-2280: Properly close secondary s2s connection on cluster join
When two servers that already have an outbound server-to-server connection to a remote domain join the same local Openfire cluster, Openfire will close one of the s2s connections (this is because Openfire currently does not support having more than one cluster node with an outbound s2s connection to the same domain - something that is subject to improvement under OF-2301).

This commit fixes an issue that prevented the second connection to be removed from the local routing table of the host that it got removed in. This bug is likely introduced by the previous commit for OF-2280.