OF-3078: Increase default JDBCAuthProvider bcrypt cost
Password hashes are easier to brute force than they would be if following the latest guidance. Reasonably strong passwords will still resit brute forcing in a reasonable time frame on current hardware.

The default (previously defined to be 10) is used only when the property `jdbcAuthProvider.bcrypt.cost` is not set. Thus, the change that is applied in this commit can be worked around (for backwards compatibility) by setting that property to the original value, 10.