OF-3083: SerializingCache should skip JAXB for simple classes
When provided with String or JID instances, full JAXB marshalling can easily be skipped (as these instances can easily be converted to and from String representations. By not requiring a JAXB conversion, this class becomes a lot more versatile.

With this change, JID and String instances can be used directly as keys and/or values for the cache. There is no need to provide a JAXB wrapper for these any longer.