OF-3083: SerializingCache to release classes when plugin unloads
This ensures that no class references are retained by the cache after a plugin (that provided the class definitions) unloads. This allows the PluginClassLoader to properly be discarded. That in turn prevents ClassCastExceptions, that will otherwise occur when a new instance of the plugin (which provides the same class definition, but loaded by a new PluginClassLoader) interacts with the cache.

To address this issue, a cache can now be 'dereferenced' (using a new method `org.jivesoftware.util.cache.CacheFactory#dereferenceSerializingCache`), which should be done whenever the plugin (that provides the class definitions) gets destroyed.

When the cache is recreated (or when the cache gets used again) by a new instance of the plugin (by invoking `org.jivesoftware.util.cache.CacheFactory#createSerializingCache`) the class definitions from the new PluginClassLoader will automatically get registered with the cache again.

When the cache gets deregistered, it becomes mostly unusable, but no data is lost.