OF-3084 OF-3087: Remove deprecated `PluginManager#getPlugin()`
The method removed by this commit depends on a name which is not under control of the plugin's author: the file name of the JAR that is used to deploy the plugin. The `getPluginByName` method provides more predictable results.

In some instances, the filename-based lookup is still required (for example, the Admin Console, which is loaded as a plugin, does not have a 'name'). For those instances, a new method is introduces (`getPluginByCanonicalName`) that is equivalent to the removed method (this effectively is a rename of that method, wrapping the result in an `Optional`). Now that the old method name is no longer available, developers are forced to explicitly choose between the two available alternatives. Generally speaking, the `getPluginByName` method should be used.