OF-3094: C2S SASL EXTERNAL must not challenge after receiving empty IR
When receiving an empty Initial Response, no challenge for an Initial Response must be issued. This seems to be caused by a simple negation bug in the code that attempts to distinguish between 'no' and 'empty' Initial Response values.

It is perfectly valid for client-to-server SASL EXTERNAL to not provide an authzid, thus an empty Initial Response is not problematic

(cherry picked from commit 39300e056a7f7b8c42f34f1f328010aaff2fabcd)