OF-3095: Adjust user idle ping cache max lifetime
Using the default / unchanged cache settings, Openfire currently logs:

> WARN [main]: org.jivesoftware.openfire.muc.spi.MultiUserChatServiceImpl - The cache that tracks Ping requests (MUC Service Pings Sent) has a maximum lifetime entry (PT30M) that is shorter than the time that we wait for responses (PT1H). This will result in (some) occupants not being removed when they fail to respond to Pings. An Openfire admin should adjust the configuration.

The default values should be configured in such a way that this warning does not need to be given.

The default maximum user idle time is 4 hours. A ping response is checked at every 25% of that time. Thus, the default cache life time should at least be an hour (but not more than 4 hours).