OF-3098: Do not kick MUC user when other resources are active
A user can join the same MUC using multiple resources. The 'ghost' detection, would completely remove the user from the MUC if only _one_ of these resources would become unresponsive.

This leads to scenarios where someone that is joined on a mobile device and a laptop to be removed when only the laptop goes into hibernation.

When a resource becomes unresponsive, only that resource is to be removed from the room. The users is kicked only after _the last_ resource leaves.