OF-2034: Add support for XEP-0398: User Avatar to vCard-Based Avatars Conversion
This adds basic support for XEP-0398, that:
- converts vCard avatars into PEP avatars
- converts PEP avatars into vCard avatars
- adds an avatar hash to presence available stanzas sent by users (per section 4)

Additionally, the code checks if vCard avatars are to be migrated when a vCard is loaded from a vCard provider. This aims to pick up changes that are applied to external data stores (such as LDAP / AD).

The existing code does not _remove_ avatars, when a removal is initiated in one of both systems.