OF-3108: Properly stop Jetty Server components
When Jetty is being stopped, individual components need to be stopped too. This prevents issues when those components are restarted again (eg: after a reconfiguration of the server).

This commit fixes what appears to be a copy/paste bug in the method that intends to stop all components (the bug caused one component to be stopped twice, while another didn't get stopped at all).