OF-3108: Properly start Jetty Server components
When Jetty is being started, individual components need to be started too.

This commit fixes what appears to be a copy/paste bug in the method that intends to start all components (the bug caused one component to be started twice, while another didn't get stopped at all).

Note that this is very, very similar to a recent commit that fixed the exact same problem when _stopping_ things...