OF-3111: Allow Jetty graceful stop
This introduces two new properties (that default to 5 seconds) that define the maximum duration that Jetty will wait for a graceful shutdown to finish. Prior to this change, Jetty was not configured to obsure such a timeout.

- webbinding (websocket/bosh) property: `httpbind.stop-timeout`
- admin console property: `adminConsole.stop-timeout`

Set to a non-positive value to disable the feature.

I'm not sure if the default value of 5 seconds is enough for HTTP Binding / BOSH requests (which is based on long-polling). Those may need a timeout equal to their long-polling duration (typically something like a minute) to be closed down gracefully. As longpolling is a practice that is being replaced by using websockets, and such a long timeout may delay a restart, I've opted to keep the default low. If you use BOSH or HTTP Binding, you may want to increase the value of the `httpbind.stop-timeout` property.