OF-3112: Relax restrictions around whitespace in Base64
in https://igniterealtime.atlassian.net/browse/OF-3085 and https://igniterealtime.atlassian.net/browse/OF-3040, a custom Base64 implementation was replaced with a Java-native implementation.

The current application is fairly strict in how it accepts whitespace characters. This causes failures (eg: with base64 encoding that has lines of 74 instead of the typical 76 characters).

XMPP specifications are rather permissive:

XEP-0153, §4.6:

> The image data MUST conform to the base64Binary datatype [7] and thus be encoded in accordance with Section 6.8 of RFC 2045 [8], which recommends that base64 data should have lines limited to at most 76 characters in length. However, any whitespace characters (e.g., '\r' and '\n') MUST be ignored.

XEP-0084, §4.1 has:

> … Base64-encoded … (Note: Line feeds SHOULD NOT be added but MUST be accepted.)

Openfire should be equally permissive.