PhotoResizer should silently ignore malformed Base64 data
When a VCard contains data that can't be processed by the PhoteResizer, it should silently ignore the data instead of erroring out. This is in line with other failures in the PhotoResizer code (e.g. when the mime type of the image isn't recognized). This will prevent clients from receiving VCard data at all, when the resizing fails.