OF-3113: Fix UserNotFoundException in PresenceSubscribeHandler
The original code seems to assume that all usernames provided in arguments reference existing users. When that's not the case, a UserNotFoundException was thrown.

Originally, the method returned 'true' only when prober and probee had a specific relation based on presence subscription. It stands to reason that when the users do not exist (Openfire does not support presense subscription to anonymous users) the resulting value should be 'false'.