OF-3113: Ensure compliance to RFC6121 Section 8.5.1
Apparently, the bug fixed in OF-3113 had a side effect that Openfire was compliant to RFC6121, with regards to the handling of certain presence stanzas that are addressed to non-existing local users.

In this commit, an explicit check is added to ensure Openfire remains compliant.

(cherry picked from commit 4c8e8bf7f1dc21b7fdc066db68ae68f9a65e0932)