OF-3115: Mark sessions as unresumable when not closed neatly
This inverts the approach of 'assume that a connection is closed correctly, unless some kind of error was detected' with 'assume that the connection is closed with some kind of error state, unless we can detect that it was closed correctly'.

It is assumed that this results in better behavior, in particular with regards to the resumability of sessions (in context of Stream Management).

To detect that a connection is closed correctly:
- see if the peer sent an 'end of stream' (typically a `</stream:stream>` tag)
- determine if close was invoked through some administrative action (eg: an admin killing a session on the admin console).

This commit:
- Stores the 'has received end-of-stream' in a boolean property of a session (although this property currently goes unused)
- Explicitly marks a session as non-resumable in many more cases
- Adds clustering support for marking a session as non-resumable (allowing an admin to properly close a session on another node)
- Enhances the `close()` API with a method that takes a Stream Error object. When not null, the session is automatically marked as non-resumable
- Applies all of the above to all types of sessions with XMPP entities. Apart from clients, this includes servers, components, and multiplexers. This facilitates Stream Management on those types of connection.