OF-3121: Unit tests for plugin (un/re)loading
This adds a new set of unit tests that verify the behavior of PluginManager's loading, unloading and deleting of plugins.

The commit introduces a set of five test plugins. The source code of these plugins is added in a TAR archive (for future maintenance).

The test plugins live in a hierarchy. Child plugins reference classes that are provided by their parents, which allows for testing of
class loading that occurs when plugins are initialized and destroyed.

This is a diagram of the hierarchy:

  Node_A
  ├── Node_A_B
  └── Node_A_C
      ├── Node_A_C_D
      └── Node_A_C_E

A README.txt file has been provided in the directory holding the test plugins to explain what they are and how they're to be maintained.