OF-3128: Immediately reload applicable parent/child plugins
When a plugin gets unloaded, all plugins in the same hierarchy get unloaded, too. This is to clean up the PluginClassLoader: it can no longer reference the deleted plugin.

Some of the other plugins in the hierarchy can continue to operate (for example, a parent or sibling plugin doesn't depend the deleted plugin to be available). These plugins are reloaded.

The process to reload these plugins is implemented by PluginMonitor. Prior to this change, this process required a second iteration of the PluginMonitor's MonitorTask for the dependent plugins to be reloaded. That is undesirable.

The loading of a plugin depends on the plugin jar file being extracted to a subdirectory.

Prior to this change, MonitorTask first extracts plugin files (that weren't extracted yet). Then, plugins that need to be unloaded were unloaded. This causes the extracted plugins from dependant plugins to be removed, too. That in turn prevents them from being installed immediately.

To have these plugins be reloaded in the same iteration as the one that deletes the original plugin, this commit changes the order in which plugins are unloaded and plugin files are extracted. By first deleting plugins, then extracting any plugins that need to be (re)loaded, the plugins can be re(loaded) immediately.