OF-3121: Ensure plugin classloader is available when destroying hierarchy
Openfire plugins can exist in a hierarchy. All plugins in the hierarchy share one PluginClassLoader.

In preexisting code, this class loader gets destroyed when the first plugin in the hierarchy gets destroyed. This is problematic when other plugins in the hierarchy require classes to be loaded when their `destroyPlugin()` method is invoked. This is relatively rare (most classes would already be loaded), but when it does occur, leaves the plugin in a broken state. Only a restart of Openfire can then resolve things.

This commit is a rewrite of the plugin loading and unloading. It reimplements how plugins in a hierarchy are iterated over, using recursion. It ensures that the PluginClassLoader is only destroyed after all plugins in the hierarchy are destroyed.

These changes will introduce some behavioral changes with regards to the order in which PluginListeners are invoked (see JM-1039 for a description of behavior that is now likely modified). I assume that this generally is an improvement, and otherwise acceptable.