OF-3131: Move 'subject' from history to room
This commit shifts responsibility for maintaining the last stanza that changed a MUC room's subject from its HistoryStrategy to the MUCRoom implementation itself.

The purpose for this change is to ensure that a MUCRoom has all data required to broadcast its latest subject.

The changes in this commit include:
- datatype change for `MUCRoom#subject` from String to Stanza
- deprecation of various subject-related methods in `HistoryStrategy`
- database migration to copy the latest subject stanza from table `ofMucConversationLog` to `ofMucRoom`

The database migration is not perfect: it does not migrate the stanza's timestamp, which means that the subject that is broadcast no longer mentions _when_ the subject was changed (as the stanzas in `ofMucConversationLog` do have an occupantJID in its 'from' attribute, the 'who' is correctly retained).

In rare occasions (see OF-3131) the room can have a subject, while the history does not. In those cases, the database migration scripts leave the (non-stanza) subject in the ofMucRoom table intact (this means that the column holds a mixture of plain text and XMPP data). The code that reads from the database will generate a stanza on the fly from such a plain text subject.