OF-3132: Improve iteration over ClientSessions of a user
Additional fixes for iterations over sessions of a user.

Broadcast data to _all_ sessions of a particular user, including those that are not yet available (fixing TODO's from @gdombiak as if its 2007!)

When iterating over one's own sessions, provide the owning JID as the 'requestor' which theoretically maybe includes sessions that have send directed presence.