Use assertThrows instead of try-catch
Follow JUnit Jupiter's best practices.
In JUnit 3 the idiomatic way for asserting an exception that's
expected was to have an unconditional `fail` statement in a try block
and additional asserts in the catch block:
```java
try {
    codeThatShouldFail();
    fail("Shouldn't have gotten here");
} catch (SomeException e) {
    assertEquals("some error message", e.getMessage());
}
```

While this does work, it's often hard to follow the flow of code written
like this, and it's relatively easy to introduce mistakes when
refactoring it.

Instead of this boilerplate, JUnit 5 allows an easier approach, using
`assertThrows`:
```java
SomeException e = assertThrows(SomeException.class, () -> codeThatShouldFail());
assertEquals("some error message", e.getMessage());
```