Offer dialback when domain is not valid.
This commit adds a configuration option, "xmpp.server.dialback.offer", that can be used to control if
dialback stream feature should be offered.

One usecase is where openfire is used as a gateway between XMPP domains ("trunking") -
https://download.igniterealtime.org/openfire/docs/latest/documentation/trunking-guide.html

In situations where dialback has to be enabled in the gateway, it also forces the server(s) behind
the gateway to use dialback. Before this commit, the code did not add dialback unless the certificates
are selfsigned. Which means that a session to a remote domain via the gateway will not be established
both ways due to lack of dialback, when the certificates are signed by CAs and trust exists.

With "xmpp.server.dialback.offer" set to "true", dialback will be added as a stream feature and
sessions can be established in both directions.