OF-3135: Improve wording of TLS Auth setting
In server-to-server configuration on the admin console, there’s a checkbox that controls if any other authentication mechanism can be used, if SASL EXTERNAL was attempted, but failed.

The rationale being that if a strong mechanism fails, using a weaker mechanism can lead to exploitable security loopholes.

The current text for this option is:

> If attempting to validate a cerificate fails, the connection is closed and not attempted via dialback authentication

This is confusing. Improve it, for example by using:

> Treat invalid TLS certificates as fatal (do not attempt other, possibly weaker, forms methods of authentication)

(cherry picked from commit 994d9b74c4275ec41d02526db8a50f28f5edd46d)