OF-3142: ServerTrafficCounter 'amount' statistics should not reset
In an earlier commit, the statistics defind by ServerTrafficCounter, which provide a rate of traffic since the last read of the statistic, have been duplicated. The duplicates are expected to provide a value that does not reset (but can increment).

By accident, the new statistics were still being reset after use. This commit fixes that.