OF-3142: Use distinct keys for multi-stats
Under OF-3142, definitions of Statistics have been modified. To help backwards compatibility (primarily for the Monitoring plugin) the old definitions are not (yet) removed.

Some Statistics are grouped in 'multi-stats'. When the duplicated and original statistics are grouped in the same 'multi-stat', graphing in the Monitoring Plugin is being confused by seeing to many data points (resulting in IndexOutOfBoundExceptions). This commit prevents that by creating distinct 'multi-stats', rather than grouping everything under the same instance.