docs: add README to doc/ folder explaining its purpose
The doc/ folder contains code-level architectural documentation (ADRs) but lacked explanation of its purpose. This was confusing as it could easily be mistaken for the documentation/ folder which serves a
different audience.

The README clarifies that doc/ is for internal architecture documentation aimed at developers working on the Openfire codebase, whilst documentation/ contains user-facing guides for administrators and plugin developers.