OF-3148: Properly persist 'admin' affiliation in MUC
By updating the in-memory affiliation of a user before determining the _old_ affiliation, the remainder of the code assumes that an _update_ rather than an insert is being executed. The corresponding database query updates zero rows, which causes the newly added affiliation to not be persisted.

(cherry picked from commit cc3c8053dddacd2237de54d18aa6a7e01fbecfab)