OF-3153: Introduce option to configure failfast behavior of room loading
This introduces a new SystemProperty `xmpp.muc.loading.failfast` that controls if an exception while loading a room causes an error to be thrown (fail-fast), or will allow all other rooms to be loaded (resilient).

Note that this applies to non-SQL exceptions only, as SQL exceptions are logged but otherwise ignored by the `loadFromDB` method.