Added comment on jansi dependency
To save future us from digging into this again: although there are later versions of Jansi, those no longer the functionality that we're using it for (`HtmlAnsiOutputStream` which is used to render the log files on the admin console). See https://github.com/fusesource/jansi/issues/210

If there's ever a need to upgrade this dependency (for example, because vulnerabilities are reported against it) we will have to replace it with a different implementation.